Refactor: Moved hamming distance to its own file
