+++ PhysiCell/custom_modules/custom.cpp
@@ -67,6 +67,7 @@
 #include <typeinfo>
 #include <vector>
 #include "./custom.h"
+#include "./hamming.h"
 
 Cell_Definition* naive_bcell;
 
@@ -210,20 +211,14 @@ void contact_function( Cell* pMe, Phenotype& phenoMe , Cell* pOther, Phenotype& 
 
 
 
-//letters not in the human amino acid alphabet: b,j,o,u,x,z
-static const double alphabet[] {'a','c','d','e','f','g','h','i','k','l','m','n','p','q','r','s','t','v','w','y'};  // humman amino acide alphabet
-static const double pad {0};
 
-// number of matching antigen antibody amino sequences that account for 100% affinity.
-static double aminoComplete = 10;
 
 void create_naive_bcell_type( void )  {
 
 	naive_bcell = find_cell_definition( "B_naive" );
 
         // antigen variable
-	//std::vector<double> antigenSequence = {pad,pad,pad,pad,pad,pad,pad,pad,pad,pad,pad,pad,pad,pad,pad,pad};
-        std::vector<double> antigenSequence {'c','a','d','d','c','e','n','k','l','l','c',pad,pad,pad,pad,pad};
+        std::vector<double> antigenSequence {'c','a','d','d','c','e','n','k','l','l','c',PAD,PAD,PAD,PAD,PAD};
 	naive_bcell->custom_data.add_vector_variable( "antigenSequence", antigenSequence );
         long antigenLength = antigenSequence.size();
         printf("number of antigenSequence ELEMENTS: %ld\n", antigenLength);
@@ -281,113 +276,4 @@ void naive_bcell_phenotype( Cell* pCell, Phenotype& phenotype , double dt ) {
 	double rApoptosis = s0Apoptosis + (s1Apoptosis - s0Apoptosis) * rPressure;
         set_single_behavior( pCell, "apoptosis" , rApoptosis );
         printf("apoptosis min: {%g}\tmax: {%g}\tset: {%g}\n", s0Apoptosis, s1Apoptosis, rApoptosis);
-}
-
-double alignment( Vector_Variable antigenSequence, Vector_Variable antibodySequence ) {
-
-    // strip and print input
-    std::vector<double> antigen_sequence {};
-    std::vector<double> antibody_sequence {};
-
-    for (double element : antigenSequence.value) {
-        if (element != pad) {
-            antigen_sequence.push_back(element);
-        }
-    }
-    for (double element : antibodySequence.value) {
-        if (element != pad) {
-            antibody_sequence.push_back(element);
-        }
-    }
-
-    printf("***antigen_sequence: ");
-    for (double element : antigen_sequence) {
-        printf("{%g}", element);
-    }
-    printf("***.\n");
-    printf("***antibody_sequence: ");
-    for (double element : antibody_sequence) {
-        printf("{%g}", element);
-    }
-    printf("***.\n");
-
-
-    // find smaller slide sequence and pad the longer one
-    std::vector<double> padded_sequence {};
-    std::vector<double> slide_sequence {};
-
-    int i_antigen = antigen_sequence.size();
-    int i_antibody = antibody_sequence.size();
-
-    if (i_antigen <= i_antibody) {
-        slide_sequence = antigen_sequence;
-        for (double element : antigen_sequence) {
-            padded_sequence.push_back(pad);
-        }
-        for (double element : antibody_sequence) {
-            padded_sequence.push_back(element);
-        }
-        for (double element : antigen_sequence) {
-            padded_sequence.push_back(pad);
-        }
-    }
-    else {
-        slide_sequence = antigen_sequence;
-        for (double element : antibody_sequence) {
-            padded_sequence.push_back(pad);
-        }
-        for (double element : antigen_sequence) {
-            padded_sequence.push_back(element);
-        }
-        for (double element : antibody_sequence) {
-            padded_sequence.push_back(pad);
-        }
-    }
-
-    // print padded sequence
-    printf("***padded_sequence: ");
-    for (double element : padded_sequence) {
-        printf("{%d}", (int) element);
-    }
-    printf("***.\n");
-    printf("***side_sequence: ");
-    for (double element : slide_sequence) {
-        printf("{%d}", (int) element);
-    }
-    printf("***.\n");
-
-
-    // get hamming distance.
-    double i_hammdist_max = 0.0;
-    int i_slide = slide_sequence.size();
-    int i_template = padded_sequence.size() - i_slide;
-
-    for (size_t i=0; i <= i_template; i++) {
-        double i_hammdist = 0.0;
-        printf("***sequence intersection: ");
-        for (size_t j=i; j < (i + i_slide); j++) {
-            if (padded_sequence[j] == slide_sequence[j-i]) {
-                i_hammdist = i_hammdist + 1.0;
-            }
-            // print slide sequences
-            printf("{%d}", (int) padded_sequence[j]);
-        }
-        printf("*** hamming distance: %g.\n", i_hammdist);
-        if (i_hammdist_max < i_hammdist) {
-            i_hammdist_max = i_hammdist;
-        }
-    }
-
-    // calcualte hammingdistance score.
-    if (i_slide < aminoComplete) {
-        printf("Warning : aminoComplete {%d} is greater than the smaller squence {%d}, hamming score can never reach 1!\n", (int) aminoComplete, i_slide);
-    }
-
-    double r_hammscore =  i_hammdist_max / aminoComplete;
-    if (r_hammscore > 1) {
-        r_hammscore = 1;
-    }
-
-    // output
-    return(r_hammscore);
 }
